Enable marching cube with INR (#1)
Enable marching cube with INR in-shader inference (base: cfb0b4e)

+++ api.cpp
@@ -308,7 +308,7 @@ int vnrNeuralVolumeGetTrainingStep(vnrVolume self)
 {
   auto nv = castNeuralVolume(self);
   nv->neural.statistics(nv->stats);
-  return nv->stats.step;
+  return (int)nv->stats.step;
 }
 
 int vnrNeuralVolumeGetNumberOfBlobs(vnrVolume self)
